% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{as.data.frame.fitted.bmgarch}
\alias{as.data.frame.fitted.bmgarch}
\title{as.data.frame method for fitted.bmgarch objects.}
\usage{
\method{as.data.frame}{fitted.bmgarch}(x, ...)
}
\arguments{
\item{x}{fitted.bmgarch object.}

\item{...}{Not used.}
}
\value{
Data frame.
}
\description{
as.data.frame method for fitted.bmgarch objects.
}
\author{
Stephen R. Martin
}
