% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{as.data.frame.forecast.bmgarch}
\alias{as.data.frame.forecast.bmgarch}
\title{as.data.frame method for forecast.bmgarch objects.}
\usage{
\method{as.data.frame}{forecast.bmgarch}(x, backcast = TRUE, ...)
}
\arguments{
\item{x}{forecast.bmgarch object.}

\item{backcast}{Logical (Default: True). Whether to include "backcasted" values from \code{\link{fitted.bmgarch}} in data frame.}

\item{...}{Not used.}
}
\value{
Data frame.
}
\description{
as.data.frame method for forecast.bmgarch objects.
}
\author{
Stephen R. Martin
}
