% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{bmgarch_list}
\alias{bmgarch_list}
\title{Collect bmgarch objects into list.}
\usage{
bmgarch_list(...)
}
\arguments{
\item{...}{bmgarch objects.}
}
\value{
List of bmgarch objects. Class: bmgarch_list and bmgarch.
}
\description{
Collect bmgarch objects into list.
}
