% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.colQTs}
\alias{.colQTs}
\title{Quantiles within lists}
\usage{
.colQTs(x, probs)
}
\arguments{
\item{x}{List of parameter sample arrays.}

\item{probs}{Quantile(s). Inherits from \code{forecast} which defaults to \code{c(.025, .975)}.}
}
\value{
Quantiles at the column level within lists
}
\description{
Obtain quantiles over columns in lists
}
\author{
philippe
}
\keyword{internal}
