% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{.colSDs}
\alias{.colSDs}
\title{Column SD's}
\usage{
.colSDs(x)
}
\arguments{
\item{x}{}
}
\value{
SD's at the columns level witin list
}
\description{
Obtain SD's over columns in lists
}
\author{
philippe
}
\keyword{internal}
