% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{.depth}
\alias{.depth}
\title{Obtain nesting depth of list}
\usage{
.depth(this, thisdepth = 0)
}
\arguments{
\item{this}{}

\item{thisdepth}{}
}
\value{
Depth
}
\description{
Figure out whther object is a list of models or just one model
Check for nesting structure; If depth == 1, one object else list of models
from: \url{https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r}
}
\keyword{internal}
