% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.extract_param_list}
\alias{.extract_param_list}
\title{Extract named parameter list from either rstan or cmdstanr fit.
Returns a named list matching the format of rstan::extract():
each element is an array [draws, dim1, dim2, ...].}
\usage{
.extract_param_list(fit, params)
}
\description{
Extract named parameter list from either rstan or cmdstanr fit.
Returns a named list matching the format of rstan::extract():
each element is an array [draws, dim1, dim2, ...].
}
\keyword{internal}
