% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.f_array_x_mat}
\alias{.f_array_x_mat}
\title{Multiply matrices in array with a vector -- generic}
\usage{
.f_array_x_mat(mat_out, array_obj, mat_obj, i)
}
\arguments{
\item{mat_out}{Output matrix.}

\item{array_obj}{3-D coefficient array.}

\item{mat_obj}{Matrix to multiply.}

\item{i}{Index of current period.}
}
\value{
matrix
}
\description{
Multiply matrices in array with a vector -- generic
}
\author{
Philippe Rast
}
\keyword{internal}
