% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.f_MA}
\alias{.f_MA}
\title{Multiply matrices in array with a vector}
\usage{
.f_MA(MA, theta, mu, rts, i)
}
\arguments{
\item{MA}{MA coefficient array.}

\item{theta}{Theta coefficient matrix.}

\item{mu}{Mean matrix.}

\item{rts}{Return time series matrix.}

\item{i}{Index of current period.}
}
\value{
matrix
}
\description{
Multiply matrices in array with a vector
}
\author{
Philippe Rast
}
\keyword{internal}
