% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdstan_path.R
\name{.get_target_stan_path}
\alias{.get_target_stan_path}
\title{Path to stan files}
\usage{
.get_target_stan_path()
}
\value{
Full path to folder containing stan files
}
\description{
Internal function to obtain stan code paths
}
\author{
philippe
}
\keyword{internal}
