% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.newline}
\alias{.newline}
\title{Print helper - Return new line(s).}
\usage{
.newline(n = 1)
}
\arguments{
\item{n}{Integer (Default: 1). Number of new lines.}
}
\value{
Prints new lines.
}
\description{
Print helper - Return new line(s).
}
\author{
Stephen R. Martin
}
\keyword{internal}
