% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.config}
\alias{.print.config}
\title{Print helper for Sampling Config.}
\usage{
.print.config(bmsum)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}
}
\value{
Void.
}
\description{
Print helper for Sampling Config.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
