% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.summary.bekk}
\alias{.print.summary.bekk}
\title{Print helper for BEKK/pdBEKK.}
\usage{
.print.summary.bekk(bmsum)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}
}
\value{
Void.
}
\description{
Print helper for BEKK/pdBEKK.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
