% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.summary.beta}
\alias{.print.summary.beta}
\title{Print helper for beta component.}
\usage{
.print.summary.beta(bmsum)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}
}
\value{
Void.
}
\description{
Print helper for beta component.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
