% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.summary.ccc}
\alias{.print.summary.ccc}
\title{Print helper for CCC.}
\usage{
.print.summary.ccc(bmsum, ...)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}

\item{...}{Not used.}

\item{x}{bmgarch.summary object.}
}
\value{
Void.
}
\description{
Print helper for CCC.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
