% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.summary.dcc}
\alias{.print.summary.dcc}
\title{Print helper for DCC.}
\usage{
.print.summary.dcc(bmsum)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}
}
\value{
Void.
}
\description{
Print helper for DCC.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
