% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.print.summary.lp}
\alias{.print.summary.lp}
\title{Print helper for LP component.}
\usage{
.print.summary.lp(bmsum)
}
\arguments{
\item{bmsum}{summary.bmgarch object.}
}
\value{
Void.
}
\description{
Print helper for LP component.
}
\author{
Stephen R. Martin, Philippe Rast
}
\keyword{internal}
