% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.qtile}
\alias{.qtile}
\title{Internal function to be used}
\usage{
.qtile(x, CrI = c(0.025, 0.975))
}
\arguments{
\item{x}{Numeric vector.}
}
\description{
Internal function to be used
}
\keyword{internal}
