% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfocv.R
\name{.refit}
\alias{.refit}
\title{Refit model}
\usage{
.refit(object, data, xC_data)
}
\arguments{
\item{object}{bmgarch model object}

\item{data}{new data}

\item{xC_data}{new predictor}
}
\description{
Refit model
}
\keyword{internal}
