% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.sep}
\alias{.sep}
\title{Print helper - Separator, new line}
\usage{
.sep()
}
\value{
Prints "---" and a new line.
}
\description{
Print helper - Separator, new line
}
\author{
Stephen R. Martin
}
\keyword{internal}
