% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.square}
\alias{.square}
\title{Internal function to be used in sweep()}
\usage{
.square(x)
}
\arguments{
\item{x}{Value to be squared}
}
\value{
Squared value
}
\description{
Internal function to be used in sweep()
}
\author{
Philippe Rast
}
\keyword{internal}
