% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{fitted.bmgarch}
\alias{fitted.bmgarch}
\title{Fitted (backcasting) method for bmgarch objects.}
\usage{
\method{fitted}{bmgarch}(object, CrI = c(0.025, 0.975), digits = 2, ...)
}
\arguments{
\item{object}{bmgarch object.}

\item{CrI}{Numeric vector (Default: \code{c(.025, .975)}). Lower and upper bound of predictive credible interval.}

\item{digits}{Integer (Default: 2, optional). Number of digits to round to when printing.}

\item{...}{Not used.}
}
\value{

}
\description{
Extracts the model-predicted means, variances, and correlations for the fitted data.
}
\details{
Whereas \code{\link{forecast.bmgarch}} computes the \emph{forecasted} values for future time periods, \code{fitted.bmgarch} computes the \emph{backcasted} (model-predicted) values for the observed time periods.
}
\author{
Stephen R. Martin
}
