% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{forecast.bmgarch}
\alias{forecast.bmgarch}
\title{Forecast method for bmgarch objects.}
\usage{
\method{forecast}{bmgarch}(object, ahead = 1, xC = NULL, CrI = c(0.025, 0.975), seed = NA)
}
\arguments{
\item{object}{bmgarch object.}

\item{ahead}{Integer (Default: 1). Periods to be forecasted ahead.}

\item{xC}{Numeric vector or matrix. Covariates(s) for the constant variance terms in C, or c. Used in a log-linear model on the constant variance terms. If vector, then it acts as a covariate for all constant variance terms. If matrix, must have columns equal to number of time series, and each column acts as a covariate for the respective time series (e.g., column 1 predicts constant variance for time series 1).}

\item{CrI}{Numeric vector (Default: \code{c(.025, .975)}). Lower and upper bound of predictive credible interval.}

\item{seed}{Integer (Optional). Specify seed for \code{\link[rstan]{sampling}}.}
}
\value{
forecast.bmgarch object. List containing \code{forecast}, \code{backcast}, and \code{meta}data. See \code{\link{fitted.bmgarch}} for information on \code{backcast}.
\code{forecast} is a list of four components. Three components (mean, var, cor) contain forecasted values in arrays. mean and var are structured as \code{[N, 7, TS]} arrays, where \code{N} is length of the time series, and \code{TS} is the number of time series. \code{cor} is structured as \code{[N, 7, TS_TS]}, where \code{TS_TS} represents a (lower triangular) correlation. All arguments are named. E.g., \code{forecasted$forecast$cor[, , "tsB_tsA"]} would the forecasted values for the correlation between the two time series, "tsB" and "tsA".
}
\description{
Estimates forecasted means, variances, and correlations from a fitted bmgarch model.
}
\author{
Stephen R. Martin
}
