% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfocv.R
\name{lfocv}
\alias{lfocv}
\title{Leave-Future-Out Cross Validation (LFO-CV)}
\usage{
lfocv(object, L = 50, mode = "backward")
}
\arguments{
\item{object}{Fitted bmgarch model. \code{lfocv} inherits all attributes from the bmgarch object.}

\item{L}{Minimal length of times series before computing LFO}

\item{mode}{forward or backward elpd_lfo approximation, or exact elpd-lfo.
Takes 'forward', 'backward', and 'exact'. 'exact' fits N-L models and may
take a \emph{very} long time to complete.}
}
\value{
Approximate LFO-CV value and log-likelihood values across (L+1):N timepoints
}
\description{
\code{lfocv} returns the LFO-CV ELPD by either computing the exact ELDP or by approximating it via
forward or backward approximation strategies based on Pareto smoothed importance sampling
described in \insertCite{Buerkner2019}{bmgarch}.
}
\references{
\insertAllCited{}
}
\author{
philippe
}
