% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{model_weights}
\alias{model_weights}
\title{Model weights}
\usage{
model_weights(
  bmgarch_objects = NULL,
  lfo_objects = NULL,
  L = NULL,
  method = "stacking"
)
}
\arguments{
\item{bmgarch_objects}{}

\item{lfo_objects}{}

\item{L}{}

\item{method}{Ensemble methods, 'stacking' (default) or 'pseudobma'}
}
\value{
Model weights
}
\description{
Model averaging via stacking of predictive distribution.
content for \description{Desc} (no empty lines) ..
.. content for \details{Det} ..
}
\author{
philippe
}
