% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panas.R
\docType{data}
\name{panas}
\alias{panas}
\title{Positive and Negative Affect Scores.}
\format{
Data frame with 200 rows and 2 variables:
\describe{
  \item{Pos}{Positive Affect score}
  \item{Neg}{Negative Affect score}
}
}
\usage{
panas
}
\description{
A dataset containing simulated values for Positive and Negative Affect scores across 200 measurement occasions for a single individual.
}
\keyword{datasets}
