% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.forecast.bmgarch}
\alias{plot.forecast.bmgarch}
\title{Plot method for forecast.bmgarch objects.}
\usage{
\method{plot}{forecast.bmgarch}(x, type = "mean", askNewPage = TRUE, last_t = 100, ...)
}
\arguments{
\item{x}{forecast.bmgarch object. See \code{\link{forecast.bmgarch}}.}

\item{type}{String (Default: "mean"). Whether to plot conditional means ("mean"), variance ("var"), or correlations ("cor").}

\item{askNewPage}{Logical (Default: True). Whether to ask for new plotting page.}

\item{last_t}{Integer (Default: 100). Only show \code{last_t} observations in plot.}

\item{...}{Not used}
}
\value{
List of ggplot objects (one per time series).
}
\description{
Plot method for forecast.bmgarch objects.
}
\author{
Stephen R. Martin
}
