% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{print.fitted.bmgarch}
\alias{print.fitted.bmgarch}
\title{Print method for fitted.bmgarch objects.}
\usage{
\method{print}{fitted.bmgarch}(x, ...)
}
\arguments{
\item{x}{fitted.bmgarch object.}

\item{...}{Not used.}
}
\value{
object (invisible).
}
\description{
Print method for fitted.bmgarch objects.
}
\author{
Stephen R. Martin
}
