% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{print.forecast.bmgarch}
\alias{print.forecast.bmgarch}
\title{Print method for forecast.bmgarch objects.}
\usage{
\method{print}{forecast.bmgarch}(x, ...)
}
\arguments{
\item{x}{forecast.bmgarch object. See \code{\link{forecast.bmgarch}}}

\item{...}{Not used.}
}
\value{
x (invisible).
}
\description{
Print method for forecast.bmgarch objects.
}
\author{
Stephen R. Martin
}
