% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfocv.R
\name{print.loo.bmgarch}
\alias{print.loo.bmgarch}
\title{print method for lfocv}
\usage{
\method{print}{loo.bmgarch}(x, ...)
}
\arguments{
\item{x}{lfo object}

\item{...}{Not used.}
}
\value{
Invisible lfocv object
}
\description{
print method for lfocv
}
\author{
philippe
}
