% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{print.model_weights}
\alias{print.model_weights}
\title{Print method for model_weights}
\usage{
\method{print}{model_weights}(x, ...)
}
\arguments{
\item{x}{Model weights object}

\item{...}{Not used.}
}
\value{
model_weights objects with weights, list of log-likelihoods, and r_eff_list
}
\description{
Print method for model_weights
}
\author{
philippe
}
