% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.bmgarch}
\alias{print.summary.bmgarch}
\title{Print method for bmgarch.summary objects.}
\usage{
\method{print}{summary.bmgarch}(x, ...)
}
\arguments{
\item{x}{summary.bmgarch object.}

\item{...}{Not used.}
}
\value{
x (invisible).
}
\description{
Print method for bmgarch.summary objects.
}
\author{
Stephen R. Martin
}
