% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch.R
\name{standat}
\alias{standat}
\title{Standardize input data to facilitate computation}
\usage{
standat(data, P, Q, ahead, standardize_data, distribution)
}
\arguments{
\item{data}{Time-series data}

\item{ahead}{Number of periods forecasting ahead}
}
\value{
bmgarch object
}
\description{
Standardize input data to facilitate computation
}
\keyword{internal}
