% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{Daily data on returns of Toyota, Nissan, and Honda stocks.}
\format{
Data frame with 2015 rows and 5 variables:
\describe{
  \item{date}{Date}
  \item{t}{Sequential time index}
  \item{toyota}{Daily returns for Toyota stock}
  \item{nissan}{Daily returns for Nissan stock}
  \item{honda}{Daily returns for Honda stock}
}
}
\usage{
stocks
}
\description{
A dataset used by Stata to illustrate MGARCH models containing daily data on returns of Toyota, Nissan, and Honda stocks.
}
\keyword{datasets}
