% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary_bmgarch}
\alias{summary_bmgarch}
\title{Summary for BMGARCH object}
\usage{
summary_bmgarch(object, CrI = c(0.05, 0.95))
}
\arguments{
\item{object}{}

\item{CrI}{}
}
\value{
Summary object
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
philippe
}
