% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch-package.R
\docType{package}
\name{bmgarch-package}
\alias{bmgarch-package}
\alias{bmgarch}
\title{The 'bmgarch' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
