% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch.R
\name{bmgarch}
\alias{bmgarch}
\title{Bayesian Multivariate GARCH}
\usage{
bmgarch(data, parameterization = "CCC", P = 1, Q = 1, ahead = 1,
  iterations = 1000, chains = 4, standardize_data = TRUE,
  distribution = "Normal", ...)
}
\arguments{
\item{data}{A time-series or matrix object containing observations at the same interval.}

\item{parameterization}{A character string specifying the type of of parameterization, must be one of "CCC" (default), "DCC", or "BEKK".}

\item{ahead}{A number specifying the number of forecasted periods. Defaults to 1.}

\item{iterations}{A positive integer specifying the number of iterations for each chain (including warmup). The default is 1000}

\item{chains}{A positive integer specifying the number of Markov chains. The default is 4.}

\item{...}{Additional arguments can be ‘chain_id’, ‘init_r’, ‘test_grad’, ‘append_samples’, ‘refresh’, ‘enable_random_init’. See the documentation in ‘stan’.}
}
\value{
An object of S4 class ‘stanfit’ representing the fitted results.
}
\description{
Draw samples from a specified multivariate GARCH model, given multivariate time-series.
}
\details{
Three paramerization are implemented. The constant conditinal correlation (CCC), the dynamic conditional correlatoin (DCC), and the BEKK.
}
\author{
philippe
}
