% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.f_array_x_mat}
\alias{.f_array_x_mat}
\title{Multiply matrices in array with a vector -- generic}
\usage{
.f_array_x_mat(mat_out, array_obj, mat_obj, i)
}
\arguments{
\item{mat_out}{}

\item{array_obj}{}

\item{mat_obj}{}

\item{i}{}
}
\value{
matrix
}
\description{
Multiply matrices in array with a vector -- generic
}
\author{
Philippe Rast
}
\keyword{internal}
