% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{.f_MA}
\alias{.f_MA}
\title{Multiply matrices in array with a vector}
\usage{
.f_MA(MA, theta, mu, rts, i)
}
\arguments{
\item{MA}{}

\item{theta}{}

\item{mu}{}

\item{rts}{}

\item{i}{}
}
\value{
matrix
}
\description{
Multiply matrices in array with a vector
}
\author{
Philippe Rast
}
\keyword{internal}
