% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{.get_stan_summary}
\alias{.get_stan_summary}
\title{Get stan summaries.}
\usage{
.get_stan_summary(object, params, CrI)
}
\arguments{
\item{object}{bmgarch object.}

\item{params}{Character vector. Names of params to pull from stan summary.}

\item{CrI}{Numeric vector (length 2).}
}
\value{
Stan summary for parameters. Columns: mean, sd, mdn, and CrIs.
}
\description{
Get stan summaries.
}
\author{
Stephen R. Martin
}
\keyword{internal}
