% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{.pred_array_to_df}
\alias{.pred_array_to_df}
\title{Convert predictive array to data.frame.}
\usage{
.pred_array_to_df(arr, type = "backcast", param = "var")
}
\arguments{
\item{arr}{Array to convert into data frame.}

\item{type}{String. "backcast" or "forecast".}

\item{param}{String. "var", "mean", or "cor".}
}
\value{
data.frame. Columns: period, type (backcast, forecast), param (var, mean, cor), TS (which time series, or which correlation for param = cor), summary columns.
}
\description{
Helper function for as.data.frame.fitted.bmgarch and as.data.frame.forecast.bmgarch. Converts predictive array to data.frame.
}
\author{
Stephen R. Martin
}
\keyword{internal}
