% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting.R
\name{forecast}
\alias{forecast}
\title{Forecasting mean and variance}
\usage{
forecast(object, ahead)
}
\arguments{
\item{object}{Fitted bmgarch object}

\item{ahead}{Periods to be forecasted ahead}
}
\value{
Forecasted mean and variance
}
\description{
Forecasting mean and variance
}
\author{
philippe
}
