% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{forecast}
\alias{forecast}
\title{Forecasting mean and variance

Forecasts the (conditional) mean and conditional variance.

\code{forecast} takes a fitted \code{bmgarch} object, and predicts the next \code{ahead} means and variances, with uncertainty.
Time-varying predictors can be included in \code{xC}.}
\usage{
forecast(
  object,
  ahead = 1,
  type = "var",
  xC_p = NULL,
  CrI = c(0.025, 0.975),
  seed = NA,
  plot = TRUE,
  last_t = 100
)
}
\arguments{
\item{object}{bmgarch object. The fitted model used for forecasting.}

\item{ahead}{Integer (Default: 1). Periods to be forecasted ahead.}

\item{type}{String (Default: "var"). Whether to plot conditional means ("mean"), variance ("var"), or correlations ("cor").}

\item{CrI}{Numeric vector (Default: \code{c(.025, .975)}). Lower and upper bound of predictive credible interval.
Possible values are .025, .05, .10, .50, .90, .95, and .975}

\item{plot}{Logical (Default: TRUE). Should forecasts be plotted.}

\item{last_t}{Integer (Default: 100). For plotting only. Only show \code{last_t} observations in plot.}

\item{xC}{Predictor for H.}
}
\value{
Forecasted conditional means, variances and correlations. Forecasts are shown in plots (unless plot = FALSE).
Conditional variance and correlation forecasts are appended to the corresponding plots with the estimates over the given period.
Mean forecasts are appended to the observed values if no meanstructure is estimated (default), otherwise they will be appended
to the esimated means.
}
\description{
Forecasting mean and variance

Forecasts the (conditional) mean and conditional variance.

\code{forecast} takes a fitted \code{bmgarch} object, and predicts the next \code{ahead} means and variances, with uncertainty.
Time-varying predictors can be included in \code{xC}.
}
\author{
philippe
}
