% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting_gq.R
\name{forecast}
\alias{forecast}
\title{%@title Forecasting mean and variance}
\usage{
forecast(
  object,
  ahead = 1,
  type = "var",
  xC_p = NULL,
  CrI = c(0.025, 0.975),
  seed = NA,
  plot = TRUE,
  last_t = 100
)
}
\arguments{
\item{object}{Fitted bmgarch object}

\item{ahead}{Periods to be forecasted ahead}

\item{type}{Plot conditional means, variances (default) or correlations. Takes arguments "mean", "var", "cor"}

\item{xC_p}{Predictor for H}

\item{CrI}{Lower and upper bound of credible interval. Default is "c( 0.025, .975)".
Possible values are .025, .05, .10, .50, .90, .95, and .975}

\item{plot}{Should forecast be plotted. Logical argument, defaults to TRUE}

\item{last_t}{For plotting: Only show last t observations in plot}
}
\value{
Forecasted conditional means, variances and correlations. Forecasts are shown in plots (unless plot = FALSE).
Conditional variance and correlation forecasts are appended to the corresponding plots with the estimates over the given period.
Mean forecasts are appended to the observed values if no meanstructure is estimated (default), otherwise they will be appended
to the esimated means.
}
\description{
%@title Forecasting mean and variance
}
\author{
philippe
}
