% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_weights.R
\name{model_weights}
\alias{model_weights}
\title{Model averaging}
\usage{
model_weights(bmgarch_objects, lfo_objects = NULL, L = NULL)
}
\arguments{
\item{bmgarch_objects}{}

\item{lfo_objects}{}

\item{L}{}
}
\value{
Model weights
}
\description{
.. content for \description{} (no empty lines) ..
.. content for \details{} ..
}
\author{
philippe
}
