% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bmgarch}
\alias{plot.bmgarch}
\title{Plot}
\usage{
\method{plot}{bmgarch}(x, type = "mean", askNewPage = TRUE, CrI = c(0.025, 0.975))
}
\arguments{
\item{x}{bmgarch object.}

\item{type}{String (Default: "var"). Whether to plot conditional means ("mean"), variance ("var"), or correlations ("cor").}

\item{askNewPage}{Logical (Default: True). Whether to ask for new plotting page.}

\item{CrI}{Numeric vector (Default: \code{c(.025, .975)}). Lower and upper bound of predictive credible interval.}
}
\value{
plot.bmgarch (Invisibly). List of plots.
}
\description{
Plot
}
\author{
Philippe Rast
}
