% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch.R
\name{standat}
\alias{standat}
\title{Standardize input data to facilitate computation}
\usage{
standat(data, xC, P, Q, standardize_data, distribution, meanstructure)
}
\arguments{
\item{data}{Time-series data}
}
\value{
bmgarch object
}
\description{
Standardize input data to facilitate computation
}
\keyword{internal}
