% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch.R
\name{standat}
\alias{standat}
\title{Standardize input data to facilitate computation}
\usage{
standat(data, xC, P, Q, standardize_data, distribution, meanstructure)
}
\arguments{
\item{data}{Time-series data}

\item{xC}{Numeric vector or matrix.}

\item{P}{Numeric.}

\item{Q}{Numeric.}

\item{standardize_data}{Logical.}

\item{distribution}{Character.}

\item{meanstructure}{Character.}
}
\value{
bmgarch stan data list.
}
\description{
Standardize input data to facilitate computation
}
\keyword{internal}
