% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.bmgarch}
\alias{summary.bmgarch}
\title{Summary for BMGARCH object}
\usage{
\method{summary}{bmgarch}(object, CrI = c(0.05, 0.95), digits = 2)
}
\arguments{
\item{object}{}

\item{CrI}{}
}
\value{
Summary object
}
\description{
Summary for BMGARCH object
}
\author{
philippe
}
