% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmgarch.R
\docType{data}
\name{supported_models}
\alias{supported_models}
\title{Models supported by bmgarch}
\format{An object of class \code{character} of length 4.}
\usage{
supported_models
}
\description{
Models supported by bmgarch
}
\author{
Stephen R. Martin
}
\keyword{internal}
